Get rid of pointless JeiDrawable file
